skibidi

+++ src/funkin/playstate.hpp
@@ -54,7 +54,7 @@ namespace funkin
         int score = 0;
         float accuracy = 100.0f;
         raylib::Vector2 cameraTarget = raylib::Vector2();
-        unsigned int totalPlayerNotes = 0;
+        size_t totalPlayerNotes = 0;
         unsigned int misses = 0;
         float health = 50.0f;
     };
